# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ComfyUI Studio API — proxy and orchestration layer for a ComfyUI server
servers:
  - url: /api
    description: Base API path
security:
  - clerkSession: []
...
paths:
...
  /comfy/validate-nodes:
    post:
      operationId: validateComfyNodes
      tags: [comfy]
      summary: Check which node class_types in a workflow JSON are missing on the ComfyUI server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateNodesInput"
      responses:
        "200":
          description: Validation result (non-blocking — informational only)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeValidation"
        "400":
          description: Invalid request
...
  /files/upload:
    post:
      operationId: uploadFile
      tags: [files]
      summary: Upload a file and forward it to ComfyUI
      parameters:
        - name: accept
          in: query
          required: false
          schema:
            type: string
          description: >-
            Optional comma-separated list of allowed file types for this upload,
            in HTML accept format (MIME types like image/png, wildcards like
            image/*, or extensions like .png). Files that don't match are
            rejected with 422.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to upload (image or audio)
      responses:
        "200":
          description: Uploaded file reference returned by ComfyUI
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadedFile"
        "400":
          description: No file provided or parse error
        "413":
          description: File too large (limit is 100 MB)
        "422":
          description: File type not allowed by the accept filter
        "502":
          description: ComfyUI upstream unreachable or returned an error
  /files/import-output:
    post:
      operationId: importOutput
      tags: [files]
      summary: Copy a generated ComfyUI output into the input directory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportOutputInput"
      responses:
        "200":
          description: Imported input file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadedFile"
        "400":
          description: Invalid output reference
        "502":
          description: ComfyUI could not be reached
...
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk browser session cookie. Sign in through the web client before calling protected endpoints.
...
  schemas:
...
    ImportOutputInput:
      type: object
      required: [filename, jobId]
      properties:
        filename:
          type: string
          minLength: 1
        jobId:
          type: number
          minimum: 1
          description: Owned completed job containing the output.
        subfolder:
          type: string
        type:
          type: string
          enum: [output]
...
    ValidateNodesInput:
      type: object
      required: [json]
      properties:
        json:
          type: string
          description: ComfyUI API-format workflow JSON as a string
    NodeValidation:
      type: object
      required: [reachable, missingNodes, totalNodes]
      properties:
        reachable:
          type: boolean
          description: Whether the ComfyUI server could be queried for its node list
        missingNodes:
          type: array
          items:
            type: string
          description: class_type values referenced in the workflow but not installed on the server
        totalNodes:
          type: number
          description: Number of distinct class_types referenced by the workflow
...
    UploadedFile:
      type: object
      required: [name, subfolder, type]
      properties:
        name:
          type: string
        subfolder:
          type: string
        type:
          type: string