# lib/api-zod/src/generated/types/importOutputInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ComfyUI Studio API — proxy and orchestration layer for a ComfyUI server
 * OpenAPI spec version: 0.1.0
 */
import type { ImportOutputInputType } from './importOutputInputType';

export interface ImportOutputInput {
  /** @minLength 1 */
  filename: string;
  /**
     * Owned completed job containing the output.
     * @minimum 1
     */
  jobId: number;
  subfolder?: string;
  type?: ImportOutputInputType;
}